refactor: check teacher rights to classroom
change teacherboard to classrooms, refactor for PR
