+++ public/app/config/channels.js
@@ -52,11 +52,11 @@ module.exports = {
   SET_USER_MODE_CHANNEL: 'user:mode:set',
   SET_SPACE_AS_FAVORITE_CHANNEL: 'user:set-space-favorite:set',
   SET_SPACE_AS_RECENT_CHANNEL: 'user:set-space-recent:set',
-  GET_CLASSROOMS_CHANNEL: 'teacher-board:classrooms:get',
-  ADD_CLASSROOM_CHANNEL: 'teacher-board:classroom:add',
-  DELETE_CLASSROOM_CHANNEL: 'teacher-board:classroom:delete',
+  GET_CLASSROOMS_CHANNEL: 'classrooms:get',
+  ADD_CLASSROOM_CHANNEL: 'classroom:add',
+  DELETE_CLASSROOM_CHANNEL: 'classroom:delete',
   SHOW_DELETE_CLASSROOM_PROMPT_CHANNEL: 'prompt:classroom:delete:show',
   RESPOND_DELETE_CLASSROOM_PROMPT_CHANNEL: 'prompt:classroom:delete:respond',
-  EDIT_CLASSROOM_CHANNEL: 'teacher-board:classroom:edit',
-  GET_CLASSROOM_CHANNEL: 'teacher-board:classroom:get',
+  EDIT_CLASSROOM_CHANNEL: 'classroom:edit',
+  GET_CLASSROOM_CHANNEL: 'classroom:get',
 };
