+++ public/app/config/errors.js
@@ -7,6 +7,7 @@ const ERROR_SPACE_ALREADY_AVAILABLE = 'ERROR_SPACE_ALREADY_AVAILABLE';
 const ERROR_DOWNLOADING_FILE = 'ERROR_DOWNLOADING_FILE';
 const ERROR_GENERAL = 'ERROR_GENERAL';
 const ERROR_DUPLICATE_CLASSROOM_NAME = 'ERROR_DUPLICATE_CLASSROOM_NAME';
+const ERROR_ACCESS_DENIED_CLASSROOM = 'ERROR_ACCESS_DENIED_CLASSROOM';
 
 module.exports = {
   ERROR_ZIP_CORRUPTED,
@@ -15,4 +16,5 @@ module.exports = {
   ERROR_DOWNLOADING_FILE,
   ERROR_GENERAL,
   ERROR_DUPLICATE_CLASSROOM_NAME,
+  ERROR_ACCESS_DENIED_CLASSROOM,
 };
