+++ public/app/listeners/getClassroom.js
@@ -1,20 +1,29 @@
 const { GET_CLASSROOM_CHANNEL } = require('../config/channels');
 const logger = require('../logger');
 const { CLASSROOMS_COLLECTION } = require('../db');
+const { ERROR_ACCESS_DENIED_CLASSROOM } = require('../config/errors');
 
-const getClassroom = (mainWindow, db) => async (event, { id }) => {
+const getClassroom = (mainWindow, db) => async (event, { id, userId }) => {
   try {
-    // todo: check teacher can access this classroom
-
     // get classroom from local db
     const classroom = db
       .get(CLASSROOMS_COLLECTION)
       .find({ id })
       .value();
-    mainWindow.webContents.send(GET_CLASSROOM_CHANNEL, classroom);
+
+    // check teacher can access this classroom
+    if (classroom.teacherId !== userId) {
+      logger.error('this user does not have access to this classroom');
+      return mainWindow.webContents.send(
+        GET_CLASSROOM_CHANNEL,
+        ERROR_ACCESS_DENIED_CLASSROOM
+      );
+    }
+
+    return mainWindow.webContents.send(GET_CLASSROOM_CHANNEL, classroom);
   } catch (err) {
     logger.error(err);
-    mainWindow.webContents.send(GET_CLASSROOM_CHANNEL, null);
+    return mainWindow.webContents.send(GET_CLASSROOM_CHANNEL, null);
   }
 };
