+++ src/components/classrooms/AddClassroomButton.js
@@ -1,45 +1,51 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import EditIcon from '@material-ui/icons/Create';
+import AddIcon from '@material-ui/icons/Add';
+import { withStyles } from '@material-ui/core';
+import Fab from '@material-ui/core/Fab';
 import { connect } from 'react-redux';
-import Tooltip from '@material-ui/core/Tooltip';
+import PropTypes from 'prop-types';
 import Button from '@material-ui/core/Button';
-import IconButton from '@material-ui/core/IconButton';
 import { withTranslation } from 'react-i18next';
 import TextField from '@material-ui/core/TextField';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import { editClassroom } from '../../actions';
+import { addClassroom } from '../../actions';
 import {
-  EDIT_CLASSROOM_BUTTON_CLASS,
-  EDIT_CLASSROOM_INPUT_ID,
-  EDIT_CLASSROOM_VALIDATE_BUTTON_ID,
-  EDIT_CLASSROOM_CANCEL_BUTTON_ID,
+  ADD_CLASSROOM_BUTTON_ID,
+  ADD_CLASSROOM_NAME_INPUT_ID,
+  ADD_CLASSROOM_VALIDATE_BUTTON_ID,
+  ADD_CLASSROOM_CANCEL_BUTTON_ID,
 } from '../../config/selectors';
 
-class EditClassroomButton extends Component {
-  state = (() => {
-    const {
-      classroom: { name },
-    } = this.props;
+const styles = theme => ({
+  fab: {
+    position: 'absolute',
+    bottom: theme.spacing(3),
+    right: theme.spacing(3),
+  },
+});
 
-    return {
-      open: false,
-      name,
-    };
-  })();
+const DIALOG_TITLE_ID = 'form-dialog-title';
 
+class AddClassroomButton extends Component {
   static propTypes = {
-    classes: PropTypes.shape({}).isRequired,
-    dispatchEditClassroom: PropTypes.func.isRequired,
-    classroom: PropTypes.shape({
-      id: PropTypes.string.isRequired,
-      name: PropTypes.string.isRequired,
+    userId: PropTypes.string.isRequired,
+    dispatchAddClassroom: PropTypes.func.isRequired,
+    classes: PropTypes.shape({
+      fab: PropTypes.string.isRequired,
     }).isRequired,
     t: PropTypes.func.isRequired,
   };
+
+  state = (() => {
+    const { t } = this.props;
+    return {
+      open: false,
+      name: t('New Classroom'),
+    };
+  })();
 
   handleClickOpen = () => {
     this.setState({ open: true });
@@ -55,11 +61,8 @@ class EditClassroomButton extends Component {
 
   handleValidate = () => {
     const { name } = this.state;
-    const {
-      dispatchEditClassroom,
-      classroom: { id },
-    } = this.props;
-    dispatchEditClassroom({ name, id });
+    const { userId, dispatchAddClassroom } = this.props;
+    dispatchAddClassroom({ name, userId });
     this.close();
   };
 
@@ -69,35 +72,35 @@ class EditClassroomButton extends Component {
   };
 
   render() {
-    const { t } = this.props;
+    const { classes, t } = this.props;
     const { name, open } = this.state;
 
-    const DIALOG_TITLE_ID = 'form-dialog-title';
     return (
       <>
-        <Tooltip title={t('Edit this classroom.')}>
-          <IconButton
-            color="inherit"
-            onClick={this.handleClickOpen}
-            className={EDIT_CLASSROOM_BUTTON_CLASS}
-          >
-            <EditIcon />
-          </IconButton>
-        </Tooltip>
+        <Fab
+          id={ADD_CLASSROOM_BUTTON_ID}
+          aria-label={t('add classroom')}
+          className={classes.fab}
+          color="primary"
+          onClick={this.handleClickOpen}
+        >
+          <AddIcon />
+        </Fab>
+
         <Dialog
           open={open}
           onClose={this.close}
           aria-labelledby={DIALOG_TITLE_ID}
         >
           <DialogTitle id={DIALOG_TITLE_ID}>
-            Edit Classroom information
+            {t('Enter a name for your new classroom')}
           </DialogTitle>
           <DialogContent>
             <TextField
-              id={EDIT_CLASSROOM_INPUT_ID}
+              id={ADD_CLASSROOM_NAME_INPUT_ID}
               autoFocus
               margin="dense"
-              label={t("Classroom's name")}
+              label={t("Classroom's Name")}
               type="text"
               fullWidth
               value={name}
@@ -108,14 +111,14 @@ class EditClassroomButton extends Component {
             <Button
               onClick={this.handleCancel}
               color="primary"
-              id={EDIT_CLASSROOM_CANCEL_BUTTON_ID}
+              id={ADD_CLASSROOM_CANCEL_BUTTON_ID}
             >
               {t('Cancel')}
             </Button>
             <Button
               onClick={this.handleValidate}
               color="primary"
-              id={EDIT_CLASSROOM_VALIDATE_BUTTON_ID}
+              id={ADD_CLASSROOM_VALIDATE_BUTTON_ID}
             >
               {t('Validate')}
             </Button>
@@ -126,15 +129,23 @@ class EditClassroomButton extends Component {
   }
 }
 
+const mapStateToProps = ({ authentication }) => ({
+  userId: authentication.getIn(['user', 'id']),
+  userMode: authentication.getIn(['user', 'settings', 'userMode']),
+});
+
 const mapDispatchToProps = {
-  dispatchEditClassroom: editClassroom,
+  dispatchAddClassroom: addClassroom,
 };
 
 const ConnectedComponent = connect(
-  null,
+  mapStateToProps,
   mapDispatchToProps
-)(EditClassroomButton);
+)(AddClassroomButton);
 
-const TranslatedComponent = withTranslation()(ConnectedComponent);
+const StyledComponent = withStyles(styles, { withTheme: true })(
+  ConnectedComponent
+);
+const TranslatedComponent = withTranslation()(StyledComponent);
 
 export default TranslatedComponent;
