+++ src/components/classrooms/EditClassroomButton.js
@@ -1,47 +1,44 @@
 import React, { Component } from 'react';
-import AddIcon from '@material-ui/icons/Add';
-import { withStyles } from '@material-ui/core';
-import Fab from '@material-ui/core/Fab';
-import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
+import EditIcon from '@material-ui/icons/Create';
+import { connect } from 'react-redux';
+import Tooltip from '@material-ui/core/Tooltip';
 import Button from '@material-ui/core/Button';
+import IconButton from '@material-ui/core/IconButton';
 import { withTranslation } from 'react-i18next';
 import TextField from '@material-ui/core/TextField';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import { addClassroom } from '../../actions';
+import { editClassroom } from '../../actions';
 import {
-  ADD_CLASSROOM_BUTTON_ID,
-  ADD_CLASSROOM_NAME_INPUT_ID,
-  ADD_CLASSROOM_VALIDATE_BUTTON_ID,
-  ADD_CLASSROOM_CANCEL_BUTTON_ID,
+  EDIT_CLASSROOM_BUTTON_CLASS,
+  EDIT_CLASSROOM_INPUT_ID,
+  EDIT_CLASSROOM_VALIDATE_BUTTON_ID,
+  EDIT_CLASSROOM_CANCEL_BUTTON_ID,
 } from '../../config/selectors';
 
-const styles = theme => ({
-  fab: {
-    position: 'absolute',
-    bottom: theme.spacing(3),
-    right: theme.spacing(3),
-  },
-});
+class EditClassroomButton extends Component {
+  state = (() => {
+    const {
+      classroom: { name },
+    } = this.props;
 
-const DIALOG_TITLE_ID = 'form-dialog-title';
+    return {
+      open: false,
+      name,
+    };
+  })();
 
-class AddClassroomButton extends Component {
   static propTypes = {
-    userId: PropTypes.string.isRequired,
-    dispatchAddClassroom: PropTypes.func.isRequired,
-    classes: PropTypes.shape({
-      fab: PropTypes.string.isRequired,
+    classes: PropTypes.shape({}).isRequired,
+    dispatchEditClassroom: PropTypes.func.isRequired,
+    classroom: PropTypes.shape({
+      id: PropTypes.string.isRequired,
+      name: PropTypes.string.isRequired,
     }).isRequired,
     t: PropTypes.func.isRequired,
-  };
-
-  state = {
-    open: false,
-    name: '',
   };
 
   handleClickOpen = () => {
@@ -58,8 +55,11 @@ class AddClassroomButton extends Component {
 
   handleValidate = () => {
     const { name } = this.state;
-    const { userId, dispatchAddClassroom } = this.props;
-    dispatchAddClassroom({ name, userId });
+    const {
+      dispatchEditClassroom,
+      classroom: { id },
+    } = this.props;
+    dispatchEditClassroom({ name, id });
     this.close();
   };
 
@@ -69,35 +69,35 @@ class AddClassroomButton extends Component {
   };
 
   render() {
-    const { classes, t } = this.props;
+    const { t } = this.props;
     const { name, open } = this.state;
 
+    const DIALOG_TITLE_ID = 'form-dialog-title';
     return (
       <>
-        <Fab
-          id={ADD_CLASSROOM_BUTTON_ID}
-          aria-label={t('add classroom')}
-          className={classes.fab}
-          color="primary"
-          onClick={this.handleClickOpen}
-        >
-          <AddIcon />
-        </Fab>
-
+        <Tooltip title={t('Edit this classroom.')}>
+          <IconButton
+            color="inherit"
+            onClick={this.handleClickOpen}
+            className={EDIT_CLASSROOM_BUTTON_CLASS}
+          >
+            <EditIcon />
+          </IconButton>
+        </Tooltip>
         <Dialog
           open={open}
           onClose={this.close}
           aria-labelledby={DIALOG_TITLE_ID}
         >
           <DialogTitle id={DIALOG_TITLE_ID}>
-            {t('Enter a name for your new classroom')}
+            {t('Edit Classroom Information')}
           </DialogTitle>
           <DialogContent>
             <TextField
-              id={ADD_CLASSROOM_NAME_INPUT_ID}
+              id={EDIT_CLASSROOM_INPUT_ID}
               autoFocus
               margin="dense"
-              label={t("Classroom's name")}
+              label={t("Classroom's Name")}
               type="text"
               fullWidth
               value={name}
@@ -108,14 +108,14 @@ class AddClassroomButton extends Component {
             <Button
               onClick={this.handleCancel}
               color="primary"
-              id={ADD_CLASSROOM_CANCEL_BUTTON_ID}
+              id={EDIT_CLASSROOM_CANCEL_BUTTON_ID}
             >
               {t('Cancel')}
             </Button>
             <Button
               onClick={this.handleValidate}
               color="primary"
-              id={ADD_CLASSROOM_VALIDATE_BUTTON_ID}
+              id={EDIT_CLASSROOM_VALIDATE_BUTTON_ID}
             >
               {t('Validate')}
             </Button>
@@ -126,23 +126,15 @@ class AddClassroomButton extends Component {
   }
 }
 
-const mapStateToProps = ({ authentication }) => ({
-  userId: authentication.getIn(['user', 'id']),
-  userMode: authentication.getIn(['user', 'settings', 'userMode']),
-});
-
 const mapDispatchToProps = {
-  dispatchAddClassroom: addClassroom,
+  dispatchEditClassroom: editClassroom,
 };
 
 const ConnectedComponent = connect(
-  mapStateToProps,
+  null,
   mapDispatchToProps
-)(AddClassroomButton);
+)(EditClassroomButton);
 
-const StyledComponent = withStyles(styles, { withTheme: true })(
-  ConnectedComponent
-);
-const TranslatedComponent = withTranslation()(StyledComponent);
+const TranslatedComponent = withTranslation()(ConnectedComponent);
 
 export default TranslatedComponent;
